+++ lib/features/food/di/food_dependency_container.dart
@@ -3,7 +3,6 @@ import 'package:zachranobed/common/data/service/delivery_service.dart';
 import 'package:zachranobed/common/data/service/entity_pairs_service.dart';
 import 'package:zachranobed/common/data/service/food_box_service.dart';
 import 'package:zachranobed/common/data/service/meal_service.dart';
-import 'package:zachranobed/common/domain/repository/delivery_repository.dart';
 import 'package:zachranobed/features/food/data/repository/firebase_food_box_repository.dart';
 import 'package:zachranobed/features/food/data/repository/firebase_offered_food_repository.dart';
 import 'package:zachranobed/features/food/domain/repository/food_box_repository.dart';
@@ -36,7 +35,6 @@ class FoodDependencyContainer {
         GetIt.I<DeliveryService>(),
         GetIt.I<MealService>(),
         GetIt.I<EntityPairService>(),
-        GetIt.I<DeliveryRepository>(),
       ),
     );